fiq_fsm: Accelerate nonperiodic split transactions
This adds non-periodic split transaction acceleration to the FIQ.
Control/Bulk split transactions are handed off to the FIQ which
results in a much lower interrupt loading, obviating the need
for nak_holdoff_enable in many cases. The NAK holdoff will still be
useful in cases of low-throughput devices where ARM performance
needs to be maximised.