xhci: add VLI_TRB_CACHE_BUG quirk
The VL805 fetches up to 4 TRBs at a time. TRB reads don't cross a 64B
boundary, and if a TRB is fetched and is not on a 64B boundary, the read
is sized up to the next 64B boundary.

However the VL805 implements a readahead prefetch for TRBs on a transfer
ring. This fetches the next 64B after any TRB read has happened. Near
the end of a ring segment, the prefetcher can read the first 64B of the
next page in physical memory and this is where the behaviour causes a
bug.

The controller does not tag reads with which endpoint they are for, so
if the start of the next page is a ring segment used by a victim
endpoint, and the victim endpoint is about to fetch TRBs from the start
of the segment, the victim endpoint will read from the prefetched data
and not perform a read to main memory.

This is fatal if fewer than 5 valid TRBs are on the ring, as is the case
with most Interrupt endpoints and Bulk OUT endpoints. On a doorbell ring
corresponding to the host writing the first TRB in the victim segment,
the victim endpoint advances past the link TRB in the previous segment
and reads the erroneously cached TRBs. The ring cycle state bit in the
cached TRBs will be invalid and the controller does not advance the
transfer for the victim endpoint. Transfers only proceed once the cached
TRBs have been invalidated through other activity and the victim
endpoint doorbell is rung again.

To prevent the prefetcher from reading past a page boundary, there must
be no TRBs in the last 64B of a page.

Add a padding parameter to struct xhci_segment and optionally set it
using the new quirk. Use the new offset to determine where to place the
Link TRB.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>